readme updated, R1/R2 answered

+++ src/controllers/users_controller.py
@@ -4,17 +4,25 @@ from models.users import User
 from schemas.user_schema import user_schema, users_schema
 from datetime import date
 from flask_jwt_extended import jwt_required, get_jwt_identity
+from main import bcrypt
 
 users = Blueprint('users', __name__, url_prefix="/users")
 
-# NEED TO ADD PUT REQUESTS FOR "UPDATE" CRUD FUNCTIONALITY
 
 # The GET routes endpoint for getting list of all users
 @users.route("/", methods=["GET"])
 @jwt_required()
 def get_users():
+    user_id = get_jwt_identity()
+    
     # get all the users from the database table
     user_list = User.query.all()
+
+    user = User.query.get(user_id)
+    # Stop the request if the user is not an admin
+    if not user.admin:
+        return abort(401, description="Unauthorised user")
+    
     # Convert the cards from the database into a JSON format and store them in result
     result = users_schema.dump(user_list)
     # return the data in JSON format
@@ -22,7 +30,16 @@ def get_users():
 
 # The GET routes endpoint for getting a single user
 @users.route("/<int:id>/", methods=["GET"])
+@jwt_required()
 def get_user(id):
+    #get the user id invoking get_jwt_identity
+    user_id = get_jwt_identity()
+    #Find it in the db
+    user = User.query.get(user_id)
+    #Make sure it is in the database
+    if not user:
+        return abort(401, description="Invalid user")
+    
     user = User.query.filter_by(id=id).first()
     #return an error if the card doesn't exist
     if not user:
@@ -34,14 +51,20 @@ def get_user(id):
 
 # The POST route endpoint
 @users.route("/", methods=["POST"])
+@jwt_required()
 def create_user():
-    # #Create a new card
+    user_id = get_jwt_identity()
+    user = User.query.get(user_id)
+    # Stop the request if the user is not an admin
+    if not user.admin:
+        return abort(401, description="Unauthorised user")
+    # Create a new user
     user_fields = user_schema.load(request.json)
 
     new_user = User()
     new_user.user_name = user_fields["user_name"]
     new_user.email = user_fields["email"]
-    new_user.password = user_fields["password"]
+    new_user.password = bcrypt.generate_password_hash(user_fields["password"]).decode("utf-8"),
     new_user.admin = user_fields["admin"]
     
     # add to the database and commit
@@ -54,7 +77,7 @@ def create_user():
 @users.route("/<int:id>/", methods=["PUT"])
 @jwt_required()
 def update_user(id):
-    # #Create a new user
+    # Create a new user
     user_fields = user_schema.load(request.json)
 
     #get the user id invoking get_jwt_identity
@@ -64,19 +87,18 @@ def update_user(id):
     #Make sure it is in the database
     if not user:
         return abort(401, description="Invalid user")
-    # Stop the request if the user is not an admin
-    if not user.admin:
-        return abort(401, description="Unauthorised user")
+    
     # find the user
     user = User.query.filter_by(id=id).first()
     #return an error if the user doesn't exist
     if not user:
         return abort(400, description= "User does not exist")
-    #update the car details with the given values
+    
+    #update the user details with the given values
     user.user_name = user_fields["user_name"]
     user.email = user_fields["email"]
-    user.password = user_fields["password"]
-    user.admin = user_fields["admin"]
+    user.password = bcrypt.generate_password_hash(user_fields["password"]).decode("utf-8")
+    user.admin = False
     
     # add to the database and commit
     db.session.commit()
@@ -86,23 +108,23 @@ def update_user(id):
 
 # Finally, we round out our CRUD resource with a DELETE method
 @users.route("/<int:id>/", methods=["DELETE"])
+@jwt_required()
 def delete_user(id):
-    # #get the user id invoking get_jwt_identity
-    # user_id = get_jwt_identity()
-    # #Find it in the db
-    # user = User.query.get(user_id)
-    # #Make sure it is in the database
-    # if not user:
-    #     return abort(401, description="Invalid user")
-    # # Stop the request if the user is not an admin
-    # if not user.admin:
-    #     return abort(401, description="Unauthorised user")
-    # find the card
+    # get the user id invoking get_jwt_identity
+    user_id = get_jwt_identity()
+    # Find it in the db
+    user = User.query.get(user_id)
+    
+    # Stop the request if the user is not an admin
+    if not user.admin:
+        return abort(401, description="Unauthorised user")
+    
+    # find the user
     user = User.query.filter_by(id=id).first()
-    #return an error if the card doesn't exist
+    #return an error if the user doesn't exist
     if not User:
         return abort(400, description= "User doesn't exist")
-    #Delete the card from the database and commit
+    #Delete the user from the database and commit
     db.session.delete(user)
     db.session.commit()
     #return the card in the response
